+++ modules/search/init.go
@@ -5,18 +5,26 @@
 package search
 
 import (
+	"infini.sh/cloud/core/security/rbac"
+	"infini.sh/coco/core"
 	"infini.sh/framework/core/api"
+	"infini.sh/framework/core/security"
 )
 
 type APIHandler struct {
 	api.Handler
 }
 
+const Category = "coco"
+const Resource = "search"
+
 func init() {
+
+	permission:=security.GetSimplePermission(Category, Resource, string(rbac.Search))
+	security.AssignPermissionsToRoles(permission,core.WidgetRole)
+
 	handler := APIHandler{}
-
-	api.HandleUIMethod(api.GET, "/query/_suggest", handler.suggest, api.RequireLogin())
-	api.HandleUIMethod(api.GET, "/query/_recommend", handler.recommend, api.RequireLogin())
-	api.HandleUIMethod(api.GET, "/query/_search", handler.search, api.RequireLogin())
-
+	api.HandleUIMethod(api.GET, "/query/_suggest", handler.suggest, api.RequirePermission(permission))
+	api.HandleUIMethod(api.GET, "/query/_recommend", handler.recommend, api.RequirePermission(permission))
+	api.HandleUIMethod(api.GET, "/query/_search", handler.search, api.RequirePermission(permission))
 }
